Introduce BatchHandler and wire both handlers in Service
